requirements.txt
numpy
pandas
opencv-python
pyside6
PyYAML
tables
scikit-image
easydict